Move the Miles builder, rollout converter and model-setup helpers into backends/miles; core/ is backend-agnostic
training/core/ was the pre-abstraction layer and still hosted the largest
backend-specific code in the tree: slime_builder.py (the Megatron/Slime
argparse builder, 640 lines) and data_converter.py (Tinker Datum ->
Slime rollout_data), each reached only through a thin adapter in
backends/miles/. Every other backend keeps that logic in its own package.

- backends/miles/builder.py IS the builder now: MilesArgumentBuilder
  (ArgumentBuilder), formerly SlimeArgumentBuilder plus its adapter.
  The adapter had a defect: MilesBackend calls
  build_args(parallelism_config=...), which the adapter swallowed into
  **kwargs and replaced with its own never-set `parallelism=None`, so a
  client's parallelism config never reached the builder. The merged
  signature accepts both spellings and an explicit num_gpus.
- backends/miles/rollout_data.py: the Datum <-> rollout_data converter
  (TinkerDataConverter), unchanged; MilesDataConverter wraps it.
- backends/miles/model_setup.py: torch_dist path resolution, SGLang
  memory fraction, TP/PP/CP/DP auto-detect and tinker:// URI mapping,
  split out of utils/model_config.py, which keeps only the generic
  helpers (GPU count, HF config, parameter estimate, architecture).
- core/ now holds validators, loss_registry, task_manager, dependencies
  only; its and utils' package docstrings say so.
- The standalone test loaders (pack-length, DP-reduction, padding,
  RL layout) point at the new modules.

Pure moves apart from the parallelism fix; protocol + unit suites 192
passed on the pod.
